Move email footer and header components
