+++ src/emails/components/EmailFooter.tsx
