+++ src/emails/components/EmailHeader.tsx
